itechsmart_supreme/requirements.txt
# iTechSmart Supreme - Python Dependencies

# Core Framework
asyncio-mqtt==0.16.1
pyyaml==6.0.1
python-dotenv==1.0.0

# Monitoring
prometheus-api-client==0.5.3
requests==2.31.0
aiohttp==3.9.1

# Network Device Management
netmiko==4.3.0
paramiko==3.4.0
telnetlib3==2.0.4

# Cloud Providers
boto3==1.34.10  # AWS
google-cloud-compute==1.15.0  # GCP
azure-mgmt-compute==30.4.0  # Azure
azure-identity==1.15.0

# Windows/Active Directory
pywinrm==0.4.3
ldap3==2.9.1

# AI/ML
openai==1.6.1  # Optional - for online AI diagnosis

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dateutil==2.8.2
cryptography==41.0.7

# Logging
colorlog==6.8.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development (optional)
black==23.12.1
flake8==6.1.0
mypy==1.7.1